Refactor DPOS-User to extend Gateway-User (#249)
* add crosschain and gateway-user base classes
* make gateway-user multisig gateway compatible
* inherit from gateway-user instead of specifying everything inside of dpos-user
